+++ riscv/insns/vnclipu_vi.h
@@ -4,9 +4,9 @@ uint64_t int_max = ~(-1ll << P.VU.vsew);
 VI_VVXI_LOOP_NARROW
 ({
   uint64_t result = vs2_u;
-  uint64_t unsigned_shift_amount = (uint64_t)(zimm5 & ((sew * 2) - 1));
+  unsigned unsigned_shift_amount = zimm5 & ((sew * 2) - 1);
   // The immediate form supports shift amounts up to 31 only.
-  unsigned_shift_amount = std::min(unsigned_shift_amount, 31ull);
+  unsigned_shift_amount = std::min(unsigned_shift_amount, 31u);
 
   // rounding
   INT_ROUNDING(result, xrm, unsigned_shift_amount);
